// src/interfaces/user.interface.ts
export interface UserInfo {
  userName?: string
  birthDate?: string
  cpfNumber?: string
  firstName?: string
  lastName?: string
  email?: string
  phone?: string
  cepNumber?: string
  street?: string
  neighbourhood?: string
  number?: string
  complement?: string
  city?: string
  state?: string
}